fix: Early dotenv
We need to call `dotenv.config()` before other files are imported to
have the .env configuration take effect in NODE_ENV=production.
